// src/app/v2/core/enums/data-history.enum.ts
export enum DataHistory {
    FARMER = 1,
    FIELD = 2
}

export enum DataHistoryType {
    REGISTERED = 1,
    EDITED = 2,
    DELETED = 3,
    DOWNLOADED = 4,
    PRINTED = 5,
    UPLOAD_GPX = 6,
    REPLACE_GPX = 7,
    GENERATE = 8
}

export enum DataHistoryModule {
    USER = 1,
    FARMER = 2,
    FIELD = 3,
    FARMER_ID_CARD = 4,
    FARMER_LIST = 5,
    GPX = 6,
    RECOMMENDATION = 7
}